requirements.txt
python-dotenv==1.0.0


langchain==0.1.14
langchain-community==0.0.32
langchain-groq==0.1.3

faiss-cpu==1.7.4
chromadb==0.4.22

pydantic==2.5.0
pydantic-settings==2.1.0

pillow==10.1.0
pytesseract==0.3.10
paddleocr==2.7.0.3

openai-whisper==20231117
librosa==0.10.0
soundfile==0.12.1

numpy==1.26.3
pandas==2.1.4
sqlite-utils==3.35.0

requests==2.31.0
python-multipart==0.0.6

torch==2.0.1
sentence-transformers==2.2.2
streamlit>=1.36,<1.53
streamlit-option-menu==0.4.0
transformers==4.30.2
huggingface-hub==0.16.4
tokenizers==0.13.3